Fixed "no site access" error
The page did not account for the user not having permission for the base website because for some reason if the site sees that the user doesnt have that permission it doesnt store the user in the session. Not completely sure if this was intentional logic or not but I left it.